Fixed double stacktrace feedback.
When an Exception occurred, the stacktrace was added as message to an
Exception for which the same happend, resulting in a double stacktrace.
The behaviour has been changed to only add a stacktrace to the feedback
when the occurred Exception has a cause. Exceptions with enough detail
no longer have a cause and Exceptions without a cause now have enough
detail.