Refactor packages [Breaks JavaLoader standalone projects]
In the following list, `io.github.pieter12345.javaloader` is abbreviated as `jl`.
- Move `jl.dependency` to `jl.core.dependency`.
- Move `jl.exceptions` to `jl.core.exceptions`.
- Move `jl.utils` to `jl.core.utils`.
- Move `jl` classes to `jl.core`.

This leaves three main packages: `jl.core`, `jl.bukkit` and `jl.standalone`.

Important note for JavaLoaderStandalone users: The `jl.JavaLoaderProject` class has been moved to `jl.core.JavaLoaderProject` and when using the `utils` package, this has to be changed to `core.utils` as well.