Java version compatibility
Wrap URI path pointing to the JavaLoader jar in File to format it properly before passing it to the Java compiler as part of the classpath. JDK 8 is more lenient with what it allows in the class path than JDK 12 and 14 (these were tested on). This commit makes JavaLoader compatible with Java versions beyond Java 8.
Fixes #8.